Move notifications bell to You page header, avatar links to profile
- Add notification bell with unread badge to /you header (between share and settings)
- Remove notifications and profile links from user drawer sidebar
- Make sidebar avatar clickable — navigates to user's profile page
- Clean up unused imports (NotificationsOutlined, Badge, ShowChartOutlined, useUnreadNotificationCount from drawer)
